+++ packages/core/src/constants/index.ts
@@ -3,9 +3,13 @@ export { erc20ABI, erc721ABI } from './abis'
 export { etherscanBlockExplorers } from './blockExplorers'
 export type { BlockExplorer, BlockExplorerName } from './blockExplorers'
 
-export { chain, allChains, defaultChains, defaultL2Chains } from './chains'
-
-export { units } from './units'
+export {
+  chain,
+  chainId,
+  allChains,
+  defaultChains,
+  defaultL2Chains,
+} from './chains'
 
 export {
   alchemyRpcUrls,
@@ -14,3 +18,5 @@ export {
   defaultInfuraId,
 } from './rpcs'
 export type { RpcProviderName } from './rpcs'
+
+export { units } from './units'
